The per-channel parameters are now arrays indexed from 0 to numberOfChannels-1, which is more natural for programmers and what Huygens actually uses.
Still, in the HRM GUI channels are enumerated from 1 to 5, to keep the final
users happy.

This change fixed some conflictive bugs and removed plenty of hacks to
deal with the fact that all arrays had a NULL element [0].

The database doesn't need any change, this fix is backwards compatible.