Support for LIF sub-images.
The contents of the on-server LIF files are explored by hucore, and an
expanded list of subimages is presented to the user as if they were
regular images. Each selected sub-image will be added as a new task to the
deconvolution queue.

This needs a (freeware) local hucore to pre-open the images and find
out the sub-images. The config file defines a

$local_huygens_core = "/usr/local/bin/hucore";

This same file will take care in the future of making the thumbnails and
other tasks. All these hucore tasks will be gathered in
scripts/hucore.tcl.

Files are stored in the database the same way as before, but the
subimages will be processed if after the file path comes the (subimage)
enclosed in brackets.