$checkMirror was renamed to $isOriginal because it is also used for other purposes here.
Implementation of image jpeg previews: some new functions.

Enabled with two different modes in hrm_config.inc.sample:
 $useThumbnails = true;
 $genThumbnails = true;

This requires Huygens 3.3.1 or to patch 3.3.0p3 as explained at
http://support.svi.nl/wiki/HH_HRMpreviews (temporary link).

A link will be included in the success notification email to a
original/restored comparison page. Currently, following this link requires to
login first, but in the future we can fix the login page so that it jumps to
the requested URL after login.