First working version of the new Tcl script, still to be thoroughly tested. By now, it is deactivated by default in hrm_config.inc, independently of the huygens version used (3.5.2 required for all functionality).
The new script reports an estimation of the end time, which in that case is 
also shown in the job_queue.php.

Fix the locked queue manager (works for me, but we'd better test it more): now 
the queue manager is not stalled during the processing of each job, and can
launch more in parallel (in the same or other servers).

Some more comments for my other recent code.