Finish up merging 'Hackathon2013' changes into feature/gc3qm
This incorporates the last leftover changes from the 'Hackathon2013' branch
that need to get merged into this branch. It was created by using
"format-patch" and "am" using the following command:

  git format-patch -k --stdout 2bc336f~1..2bc336f | git am -3 -k --exclude=inc/Job.inc.php

Excluding the changes to "inc/Job.inc.php" was necessary as those were already
merged into our branch in commit:4c43602.

Fixes #257.

-> The branch 'Hackathon2013' should now be fully processed (cherry-picked and
hand-merged into 'devel' and 'feature/gc3qm') and thus ready for being closed.

--
Original commit message of 2bc336f:
--

(1) The Huygens template is now created when the user launches the job instead
of when the job reaches the first position of the queue.
(2) Renamed all strings refering to the "script" by strings refering to the
"template".
(3) Replaced the old .tcl script extension by the .hgsb of the Huygens Batch
templates.