Only use "sudo" in the user dir admin tool if required.
It is not necessary to use sudo for the file operations if no "SUSER" has been
set in the hrm.conf file or if the user calling the "hrm" script is already the
same that was configured in this variable. Using sudo even though the user is
already the same results in non-working commands and thus no user directories
being created.

refs #59, #173, #183, #209