Request *regular* instead of *archived* images from OMERO.
Adjust the call to the OMERO CLI so that it doesn't request the "archived"
version of an image as this was a very strict limitation (only if the user
manually ticked the "archive" option during upload, the image was kept in its
original file format). We anyway work on the OME-TIFFs provided by OMERO by
default, so there is no need for this limitation.

refs #232, #108