Added a set_time_limit() of 10 seconds to the Mail::send() method.
In case of a misconfigured mail system, the HRM might get stuck for a long time waiting for the email to be sent. A time limit of 10 seconds was now added. This will be ignored on PHP 5.3 and older if PHP was compiled with safe mode on, but should work as of PHP 5.4 since safe mode was removed.