Added a temporary configuration variable $change_ownership (that defaults to false) to enable chown/chmod in the code.
All code that explicitly changes ownership and permissions in the HRM will be removed in version 3.3. In version 3.2, this code is disabled and obsoleted, together with installation method 1. In this transition phase, we allow users to restore it by setting a configuration variable $change_ownership = true in the configuration files.