Add help links for SNR estimation tool and remove quick help for SNR page.
The top left navigation link is customized for the SNR estimator tool and the help button for the estimator algorithm now points to the SVI wiki. Moreover, the quick help for this page is removed since the static one was almost exhaustive. One sentence was added to discuss the algorithm selection.