Use string ID's containing the object class.
This is done in preparation for on-demand loading of the tree: as jqTree only
submits the ID as a parameter to the request, we need to be able to derive from
there what part of the tree to load/generate. By putting the object class into
the id we can simply parse out this information in a later call.

refs #387