Show OMERO tree by default if connection was established before.
Previously, if the user navigated away from the file manager, the OMERO
tree was not shown anymore upon return to this page, until the OMERO button
was clicked again.

As we can easily derive if a connection to an OMERO server was established
before, we can simply show the tree if such a connection exists.

refs #388