Fix SSL issue for OMERO versions >=5.4.10
On Debian installations and the above OMERO version, the push from HRM to OMERO does not work because of a failed SSL handshake.
This commit is the workaround for it.
https://forum.image.sc/t/unable-to-use-cli-importer/26424

In any case the upgrade check does not make a lot of sense in this scenario,
given that CLI output is not seen by a user nor does it currently appear in any log.