When forking the background process wait for the PID file to show up
This fixes #718 which was triggered by systemd calling the queuemanager script
which returned immediately after forking itself into the background - but the
PID file would only be created by the forked background process, which will be
a few moments too late and hence make systemd complain (as it immediately checks
for the file once the called process returns).

Simply waiting for the PID file *before* calling `exit` prevents this isue.